requirements.txt
asgiref==3.6.0
astroid==2.15.0
certifi==2022.12.7
charset-normalizer==3.1.0
dill==0.3.6
Django==4.1.7
djangorestframework==3.14.0
drf-nested-routers==0.93.4
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mysqlclient==2.1.1
platformdirs==3.1.0
pylint==2.16.4
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.5.0
urllib3==1.26.15
wrapt==1.15.0
